requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pdf2image==1.16.3
PyPDF2==3.0.1
Pillow==10.1.0
numpy<2.0.0
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0
